Fix: search-query bug on initial page load
